Moved logger to utils
