+++ api/src/routes/account.js
@@ -17,7 +17,8 @@
 import { Router } from "express";
 
 import Models from "../models/index.js";
-import { logger, mailer, ACCESS_LEVEL } from "../common_items.js";
+import { mailer, ACCESS_LEVEL } from "../common_items.js";
+import { logger } from "../utils/logging.js";
 import { committee_id_to_display } from "../db_loaded_items.js";
 
 import bcrypt from "bcrypt";
