+++ api/src/routes/login.js
@@ -19,7 +19,8 @@ import crypto from "crypto";
 import bcrypt from "bcrypt";
 
 import Models from "../models/index.js";
-import { mailer, logger, ACCESS_LEVEL } from "../common_items.js";
+import { mailer, ACCESS_LEVEL } from "../common_items.js";
+import { logger } from "../utils/logging.js";
 
 const router = Router();
 const bcrypt_rounds = 10;
