+++ api/src/server.js
@@ -26,7 +26,8 @@ import rateLimit from "express-rate-limit";
 // Import files
 import app from "./app.js";
 import { db_conn, db_check } from "./models/index.js";
-import { logger, smtp_check } from "./common_items.js";
+import { smtp_check } from "./common_items.js";
+import { logger } from "./utils/logging.js";
 import { loader_check } from "./db_loaded_items.js";
 import checkAPI from "./middleware/checkAPI.js";
 import apiLogger from "./middleware/logging.js";
